src/backend/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
python-dotenv==1.0.0
PyJWT==2.8.0
Werkzeug==2.3.7 